# Improve speed of `predictions()`