--- a/DESCRIPTION
+++ b/DESCRIPTION
@@ -1,6 +1,6 @@
 Package: marginaleffects
 Title: Predictions, Comparisons, Slopes, Marginal Means, and Hypothesis Tests
-Version: 0.9.0.9002
+Version: 0.9.0.9003
 Authors@R: 
     c(person(given = "Vincent",
              family = "Arel-Bundock",