--- a/R/datagrid.R
+++ b/R/datagrid.R
@@ -166,7 +166,7 @@ counterfactual <- function(..., model = NULL, newdata = NULL) {
         setcolorder(dat_automatic, c("rowidcf", setdiff(names(dat_automatic), "rowidcf")))
         # cross-join 2 data.tables, faster than merging two dataframes
         # https://stackoverflow.com/questions/75361701/merging-two-data-tables-that-dont-have-common-columns
-        out <- setkey(dat_automatic[, c(k=1, .SD)], k)[at[, c(k = 1, .SD)], allow.cartesian = TRUE][, k := NULL]
+        out <- cjdt(list(dat_automatic, at))
     }  else {
         out <- merge(rowid, at, all = TRUE)
     }