Benchmark:
``` r
library(marginaleffects)

tmp <- list()
for (i in 1:30000) {
  tmp[[i]] <- mtcars
}
tmp2 <- data.table::rbindlist(tmp)
tmp2$am <- as.logical(tmp2$am)
dim(tmp2)
#> [1] 960000     11

mod <- lm(mpg ~ hp + factor(cyl), data = tmp2)
```
Old:
```r
bench::mark(
  predictions(mod, variables = list(hp = c(100, 120))),
  predictions(mod, newdata = datagridcf(hp = c(100, 120), cyl = c(4, 6))),
  check = FALSE
)
#> Warning: Some expressions had a GC in every iteration; so filtering is
#> disabled.
#> # A tibble: 2 × 6
#>   expression                                                                 min
#>   <bch:expr>                                                              <bch:>
#> 1 predictions(mod, variables = list(hp = c(100, 120)))                     12.4s
#> 2 predictions(mod, newdata = datagridcf(hp = c(100, 120), cyl = c(4, 6)))  21.2s
#> # … with 4 more variables: median <bch:tm>, `itr/sec` <dbl>,
#> #   mem_alloc <bch:byt>, `gc/sec` <dbl>
```
New:
``` r
bench::mark(
  predictions(mod, variables = list(hp = c(100, 120))),
  predictions(mod, newdata = datagridcf(hp = c(100, 120), cyl = c(4, 6))),
  check = FALSE
)
#> Warning: Some expressions had a GC in every iteration; so filtering is
#> disabled.
#> # A tibble: 2 × 6
#>   expression                                                                 min
#>   <bch:expr>                                                              <bch:>
#> 1 predictions(mod, variables = list(hp = c(100, 120)))                     9.08s
#> 2 predictions(mod, newdata = datagridcf(hp = c(100, 120), cyl = c(4, 6)))  8.23s
#> # … with 4 more variables: median <bch:tm>, `itr/sec` <dbl>,
#> #   mem_alloc <bch:byt>, `gc/sec` <dbl>
```