@vincentarelbundock 2023-02-06
Very cool!

Have you tried the `cjdt()` function here? https://github.com/vincentarelbundock/marginaleffects/blob/main/R/utils.R#L62