@etiennebacher 2023-02-06
It's a 3-4 seconds slower for me (but I did only 1 iteration so could be just because of noise). Depends if you prefer saving these seconds or make the code more readable