@etiennebacher 2023-02-06
Do you know why the tests are cancelled? One failed because it couldn't download Ubuntu, did it trigger the cancellation for the rest?