@vincentarelbundock 2023-02-06
> It's 3-4 seconds slower for me (but I did only 1 iteration so could be just because of noise). Depends if you prefer saving these seconds or make the code more readable

Cool thanks. Let's keep the faster version, as long as it passes tests. Maybe just include a comment to say that this is faster than the equivalent `cjdt`.

> Do you know why the tests are cancelled? One failed because it couldn't download Ubuntu, did it trigger the cancellation for the rest?

Yes, that's my guess. I've noticed that Github seems to have changed their policy (or something in my repo settings). When one machine of a single "matrix" fails, the others are cancelled.