@etiennebacher 2023-02-06
Using `cjdt()`:
``` r
bench::mark(
  predictions(mod, variables = list(hp = c(100, 120))),
  predictions(mod, newdata = datagridcf(hp = c(100, 120), cyl = c(4, 6))),
  check = FALSE,
  iterations = 10
)
#> Warning: Some expressions had a GC in every iteration; so filtering is
#> disabled.
#> # A tibble: 2 × 6
#>   expression                                                                 min
#>   <bch:expr>                                                              <bch:>
#> 1 predictions(mod, variables = list(hp = c(100, 120)))                     5.97s
#> 2 predictions(mod, newdata = datagridcf(hp = c(100, 120), cyl = c(4, 6)))  6.64s
#> # … with 4 more variables: median <bch:tm>, `itr/sec` <dbl>,
#> #   mem_alloc <bch:byt>, `gc/sec` <dbl>
```
Cool trick!