@etiennebacher 2023-02-06
> Wait, so `cjdt()` is faster after all ?!

I think so, I'm re-running the first benchmark with several iterations to confirm