@etiennebacher 2023-02-06
Timings are very close (a few milliseconds) so let's keep `cjdt()` that is easier to read